store files in cloud

+++ server/routes/admin.js
@@ -3,8 +3,8 @@ import jwt from "jsonwebtoken";
 import VacancyModel from "../models/vacancy.js";
 import ProfileModel from "../models/profile.js";
 import FileModel from "../models/file.js";
+import { megaStorage, deleteFileById } from "../services/megaStorageEngine.js";
 import ApplicationsModel from "../models/applications.js";
-import mongoose from "mongoose";
 
 const router = express.Router({ mergeParams: true });
 
@@ -86,7 +86,11 @@ router.delete("/profile", async (req, res) => {
     try {
         const { id } = req.body;
         const { resumeId } = await ProfileModel.findById(id);
-        await FileModel.findByIdAndDelete(resumeId);
+        const isFileDeleted = await deleteFileById(resumeId);
+        if(!isFileDeleted){
+            console.log(`Failed to delete resume file of Profile with id:${id}`);
+            return res.status(500).json({error: "Failed to delete File!!"});
+        }
         await ProfileModel.findByIdAndDelete(id);
         console.log(`Profile with id:${id} is deleted successfully`);
         res.status(200).json({ message: "Delete user profile successfully!!" });
@@ -112,9 +116,20 @@ router.delete("/applications", async (req, res) => {
     try {
         const { id } = req.body;
         const { photoId, resumeId, marksheetId, signatureId } = await ApplicationsModel.findById(id);
-        await FileModel.deleteMany({ _id: [photoId, resumeId, marksheetId, signatureId] });
+        const isAllFileDeleted = await [photoId, resumeId, marksheetId, signatureId]
+            .map((id) => async (preStatus) => {
+                const isDeleted = await deleteFileById(id);
+                return isDeleted && preStatus;
+            })
+            .reduce(async (acc, fn)=>{
+                return acc.then(fn);
+        }, Promise.resolve(true));
+        if(!isAllFileDeleted){
+            console.log(`Failed to delete some files of Application with id:${id}!!`);
+            return res.status(500).json({error: "Failed to delete some files!!"});
+        } 
         await ApplicationsModel.findByIdAndDelete(id);
-        console.log(`Application with id:${id} is deleted successfully`);
+        console.log(`Application with id:${id} is deleted successfully!!`);
         res.status(200).json({ message: "Applications is deleted successfully!!" });
     }
     catch (err) {
@@ -126,14 +141,13 @@ router.delete("/applications", async (req, res) => {
 router.get("/file/:id", async (req, res) => {
     try {
         const fileId = req.params.id;
-        const readStream = FileModel.findById(fileId).cursor({
-            transform: (doc) => {
-                return doc.bufferData;
-            }
-        });
+        const { nodeId } = await FileModel.findById(fileId, "nodeId");
+        const targetFile = megaStorage.navigate("uploads").find(file => file.nodeId == nodeId);
+        if(!targetFile) return res.status(400).json({error: "File not found!!"});
+        const readStream = targetFile.download();
         readStream.on("error", async (err) => {
             console.error(err);
-            await readStream.close();
+            // await readStream.close();
             return res.status(500).json({ error: "Failed to server requested file!!" });
         });
         readStream.pipe(res);
